Add support for extra.d directory
To allow muliple plugins, sorted plugin load and general extensibility
of the web client, the server now supports to load an extra.d folder on
startup.

Sub folders inside extra.d, are loaded in alphabetic order and searched
for head.html and body.html, making it possible to append additional
html code into the <head> and <body> elements for the web client.

Example:
```
	extra.d/
	└── my-plugin
	    ├── body.html
	    ├── head.html
	    └── static
	        ├── css
	        │   └── my-plugin.css
	        └── js
	            └── my-plugin.js
```

Example head.html:
```
	<link rel="stylesheet" href="<%.S%>/css/my-plugin.css">
```

Example body.html:
```
	<script data-plugin="<%.S%>/js/my-plugin.js"></script>
```

This makes it possible to extend the web client with additional
components and/or overwrite existing.