Add script to block UDP for TURN testing
Use this script for TURN testing. Make sure Spreed WebRTC has a TURN server configured which also supports tcp. Make a call and check chrome://webrtc-internals or about:webrtc to see if TURN is actually in use (also make sure that audio and video is coming through in both directions).

https://github.com/strukturag/docker-webrtc-turnserver/issues/5